filter test cases for the old JDBC driver
