+++ src/test/java/net/snowflake/client/core/arrow/BaseConverterTest.java
@@ -1,3 +1,6 @@
+/*
+ * Copyright (c) 2012-2020 Snowflake Computing Inc. All right reserved.
+ */
 package net.snowflake.client.core.arrow;
 
 import java.util.TimeZone;
