+++ src/test/java/net/snowflake/client/jdbc/ResultJsonParserV2Test.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012-2019 Snowflake Computing Inc. All rights reserved.
+ * Copyright (c) 2012-2020 Snowflake Computing Inc. All rights reserved.
  */
 package net.snowflake.client.jdbc;
 
@@ -15,8 +15,7 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
 /** This is the unit tests for ResultJsonParserV2 */
-@Category(TestCategoryResultSet.class)
-public class ResultJsonParserV2IT {
+public class ResultJsonParserV2Test {
   @Test
   public void simpleTest() throws SnowflakeSQLException {
     SFSession session = null;
