+++ src/test/java/net/snowflake/client/jdbc/ResultSetLatestIT.java
@@ -1,0 +1,220 @@
+package net.snowflake.client.jdbc;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import net.snowflake.client.category.TestCategoryResultSet;
+import net.snowflake.client.jdbc.telemetry.*;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.experimental.categories.Category;
+
+import java.io.IOException;
+import java.sql.*;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.regex.Pattern;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.Assert.*;
+import static org.junit.Assert.assertNull;
+
+@Category(TestCategoryResultSet.class)
+public class ResultSetLatestIT extends BaseJDBCTest {
+  @Before
+  public void setUp() throws SQLException {
+    Connection con = getConnection();
+
+    // TEST_RS
+    con.createStatement().execute("create or replace table test_rs (colA string)");
+    con.createStatement().execute("insert into test_rs values('rowOne')");
+    con.createStatement().execute("insert into test_rs values('rowTwo')");
+    con.createStatement().execute("insert into test_rs values('rowThree')");
+
+    // ORDERS_JDBC
+    Statement statement = con.createStatement();
+    statement.execute(
+        "create or replace table orders_jdbc"
+            + "(C1 STRING NOT NULL COMMENT 'JDBC', "
+            + "C2 STRING, C3 STRING, C4 STRING, C5 STRING, C6 STRING, "
+            + "C7 STRING, C8 STRING, C9 STRING) "
+            + "stage_file_format = (field_delimiter='|' "
+            + "error_on_column_count_mismatch=false)");
+    // put files
+    assertTrue(
+        "Failed to put a file",
+        statement.execute(
+            "PUT file://" + getFullPathFileInResource(TEST_DATA_FILE) + " @%orders_jdbc"));
+    assertTrue(
+        "Failed to put a file",
+        statement.execute(
+            "PUT file://" + getFullPathFileInResource(TEST_DATA_FILE_2) + " @%orders_jdbc"));
+
+    int numRows = statement.executeUpdate("copy into orders_jdbc");
+
+    assertEquals("Unexpected number of rows copied: " + numRows, 73, numRows);
+
+    con.close();
+  }
+
+  @After
+  public void tearDown() throws SQLException {
+    Connection con = getConnection();
+    con.createStatement().execute("drop table if exists orders_jdbc");
+    con.createStatement().execute("drop table if exists test_rs");
+    con.close();
+  }
+
+  @Test
+  public void testMetadataAPIMetricCollection() throws SQLException {
+    Connection con = getConnection();
+    Telemetry telemetry =
+        con.unwrap(SnowflakeConnectionV1.class).getSfSession().getTelemetryClient();
+    DatabaseMetaData metadata = con.getMetaData();
+    // Call one of the DatabaseMetadata API functions but for simplicity, ensure returned ResultSet
+    // is empty
+    metadata.getColumns("fakecatalog", "fakeschema", null, null);
+    LinkedList<TelemetryData> logs = ((TelemetryClient) telemetry).logBuffer();
+    // No result set has been downloaded from server so no chunk downloader metrics have been
+    // collected
+    // Logs should contain 1 item: the data about the getColumns() parameters
+    assertEquals(logs.size(), 1);
+    // Assert the log is of type client_metadata_api_metrics
+    assertEquals(
+        logs.get(0).getMessage().get(TelemetryUtil.TYPE).textValue(),
+        TelemetryField.METADATA_METRICS.toString());
+    // Assert function name and params match and that query id exists
+    assertEquals(logs.get(0).getMessage().get("function_name").textValue(), "getColumns");
+    assertTrue(
+        Pattern.matches(
+            "[a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{12}",
+            logs.get(0).getMessage().get("query_id").textValue()));
+    JsonNode parameterValues = logs.get(0).getMessage().get("function_parameters");
+    assertEquals(parameterValues.get("catalog").textValue(), "fakecatalog");
+    assertEquals(parameterValues.get("schema").textValue(), "fakeschema");
+    assertNull(parameterValues.get("general_name_pattern").textValue());
+    assertNull(parameterValues.get("specific_name_pattern").textValue());
+
+    // send data to clear log for next test
+    telemetry.sendBatchAsync();
+
+    String catalog = con.getCatalog();
+    String schema = con.getSchema();
+    ResultSet rs = metadata.getColumns(catalog, schema, null, null);
+    logs = ((TelemetryClient) telemetry).logBuffer();
+    assertEquals(logs.size(), 2);
+    // first item in log buffer is metrics on time to consume first result set chunk
+    assertEquals(
+        logs.get(0).getMessage().get(TelemetryUtil.TYPE).textValue(),
+        TelemetryField.TIME_CONSUME_FIRST_RESULT.toString());
+    // second item in log buffer is metrics on getProcedureColumns() parameters
+    // Assert the log is of type client_metadata_api_metrics
+    assertEquals(
+        logs.get(1).getMessage().get(TelemetryUtil.TYPE).textValue(),
+        TelemetryField.METADATA_METRICS.toString());
+    // Assert function name and params match and that query id exists
+    assertEquals(logs.get(1).getMessage().get("function_name").textValue(), "getColumns");
+    assertTrue(
+        Pattern.matches(
+            "[a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{12}",
+            logs.get(1).getMessage().get("query_id").textValue()));
+    parameterValues = logs.get(1).getMessage().get("function_parameters");
+    assertEquals(parameterValues.get("catalog").textValue(), catalog);
+    assertEquals(parameterValues.get("schema").textValue(), schema);
+    assertNull(parameterValues.get("general_name_pattern").textValue());
+    assertNull(parameterValues.get("specific_name_pattern").textValue());
+  }
+
+  @Test
+  public void testMultipleChunks() throws SQLException, IOException {
+    Connection con = getConnection();
+    Statement statement = con.createStatement();
+
+    // 10000 rows should be enough to force result into multiple chunks
+    ResultSet resultSet =
+        statement.executeQuery(
+            "select seq8(), randstr(1000, random()) from table(generator(rowcount => 10000))");
+    int cnt = 0;
+    while (resultSet.next()) {
+      ++cnt;
+    }
+    assertTrue(cnt >= 0);
+    Telemetry telemetry =
+        con.unwrap(SnowflakeConnectionV1.class).getSfSession().getTelemetryClient();
+    LinkedList<TelemetryData> logs = ((TelemetryClient) telemetry).logBuffer();
+
+    // there should be a log for each of the following fields
+    TelemetryField[] expectedFields = {
+      TelemetryField.TIME_CONSUME_FIRST_RESULT, TelemetryField.TIME_CONSUME_LAST_RESULT,
+      TelemetryField.TIME_WAITING_FOR_CHUNKS, TelemetryField.TIME_DOWNLOADING_CHUNKS,
+      TelemetryField.TIME_PARSING_CHUNKS
+    };
+    boolean[] succeeded = new boolean[expectedFields.length];
+
+    for (int i = 0; i < expectedFields.length; i++) {
+      succeeded[i] = false;
+      for (TelemetryData log : logs) {
+        if (log.getMessage().get(TelemetryUtil.TYPE).textValue().equals(expectedFields[i].field)) {
+          succeeded[i] = true;
+          break;
+        }
+      }
+    }
+
+    for (int i = 0; i < expectedFields.length; i++) {
+      assertThat(
+          String.format("%s field not found in telemetry logs\n", expectedFields[i].field),
+          succeeded[i]);
+    }
+    telemetry.sendBatchAsync();
+  }
+
+  @Test
+  public void testResultSetMetadata() throws SQLException {
+    Connection connection = getConnection();
+    final Map<String, String> params = getConnectionParameters();
+    Statement statement = connection.createStatement();
+
+    statement.execute("create or replace table test_rsmd(colA number(20, 5), colB string)");
+    statement.execute("insert into test_rsmd values(1.00, 'str'),(2.00, 'str2')");
+    ResultSet resultSet = statement.executeQuery("select * from test_rsmd");
+    ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
+    assertEquals(
+        params.get("database").toUpperCase(), resultSetMetaData.getCatalogName(1).toUpperCase());
+    assertEquals(
+        params.get("schema").toUpperCase(), resultSetMetaData.getSchemaName(1).toUpperCase());
+    assertEquals("TEST_RSMD", resultSetMetaData.getTableName(1));
+    assertEquals(String.class.getName(), resultSetMetaData.getColumnClassName(2));
+    assertEquals(2, resultSetMetaData.getColumnCount());
+    assertEquals(22, resultSetMetaData.getColumnDisplaySize(1));
+    assertEquals("COLA", resultSetMetaData.getColumnLabel(1));
+    assertEquals("COLA", resultSetMetaData.getColumnName(1));
+    assertEquals(3, resultSetMetaData.getColumnType(1));
+    assertEquals("NUMBER", resultSetMetaData.getColumnTypeName(1));
+    assertEquals(20, resultSetMetaData.getPrecision(1));
+    assertEquals(5, resultSetMetaData.getScale(1));
+    assertFalse(resultSetMetaData.isAutoIncrement(1));
+    assertFalse(resultSetMetaData.isCaseSensitive(1));
+    assertFalse(resultSetMetaData.isCurrency(1));
+    assertFalse(resultSetMetaData.isDefinitelyWritable(1));
+    assertEquals(ResultSetMetaData.columnNullable, resultSetMetaData.isNullable(1));
+    assertTrue(resultSetMetaData.isReadOnly(1));
+    assertTrue(resultSetMetaData.isSearchable(1));
+    assertTrue(resultSetMetaData.isSigned(1));
+    SnowflakeResultSetMetaData secretMetaData =
+        resultSetMetaData.unwrap(SnowflakeResultSetMetaData.class);
+    List<String> colNames = secretMetaData.getColumnNames();
+    assertEquals("COLA", colNames.get(0));
+    assertEquals("COLB", colNames.get(1));
+    assertEquals(Types.DECIMAL, secretMetaData.getInternalColumnType(1));
+    assertEquals(Types.VARCHAR, secretMetaData.getInternalColumnType(2));
+    assertTrue(
+        Pattern.matches(
+            "[a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{12}",
+            secretMetaData.getQueryID()));
+
+    statement.execute("drop table if exists test_rsmd");
+    statement.close();
+    connection.close();
+  }
+}
