+++ src/test/java/net/snowflake/client/jdbc/telemetry/TelemetryTest.java
@@ -1,3 +1,6 @@
+/*
+ * Copyright (c) 2012-2020 Snowflake Computing Inc. All right reserved.
+ */
 package net.snowflake.client.jdbc.telemetry;
 
 import static org.junit.Assert.assertEquals;
@@ -8,6 +11,7 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.util.LinkedList;
 import org.junit.Test;
 
+/** Telemetry unit tests */
 public class TelemetryTest {
   private ObjectMapper mapper = new ObjectMapper();
